+++ primefaces/src/main/java/org/primefaces/component/avatar/Avatar.java
@@ -35,5 +35,7 @@ public class Avatar extends AvatarBase {
     public static final String SIZE_XLARGE_CLASS = "ui-avatar-xl";
     public static final String SIZE_TEXT_CLASS = "ui-avatar-text";
     public static final String DYNAMIC_COLOR_CLASS = "ui-avatar-dynamic";
+    public static final String DYNAMIC_COLOR_DARK_CLASS = "ui-avatar-dark";
+    public static final String DYNAMIC_COLOR_LIGHT_CLASS = "ui-avatar-light";
     public static final String ICON_CLASS = "ui-avatar-icon";
 }