+++ docs/11_0_0/gettingstarted/whatsnew.md
@@ -5,6 +5,7 @@ This page contains a list of big features. Please check the GitHub issues for al
 ## 11.0.0
 
   * AutoComplete: added `footer` facet which will be added to the suggestion list.
+  * Avatar: added `saturation`, `lightness`, `alpha` to control dynamic colors.
   * Hotkey: added `bindMac` to use alternative binding on macOS.
   * Inplace: added `tabindex` for keyboard focus and toggle support.
   * InputText, InputTextarea: counter can count bytes instead of characters.